Match slider size and integer display for clock custom width

+++ Models/Settings.cs
@@ -35,7 +35,7 @@ public sealed class Settings
     public string? ClockTimeFormatHover { get; set; }
     public string? ClockDateFormatHover { get; set; }
     public ClockWidthMode ClockWidthMode { get; set; } = ClockWidthMode.Auto;
-    public double ClockCustomWidth { get; set; } = 640;
+    public int ClockCustomWidth { get; set; } = 640;
     public Dictionary<string, bool> Visibility { get; set; } = [];
     public bool UpdateCheckEnabled { get; set; } = true;
     public UpdateCheckFrequency UpdateFrequency { get; set; } = UpdateCheckFrequency.Daily;
