+++ ViewModels/SettingChange.cs
@@ -35,5 +35,5 @@ public abstract record SettingChange
 
     public sealed record WidgetStyle(string Widget, string Family, int Scale, WidgetFontWeight Weight) : SettingChange;
 
-    public sealed record ClockWidth(ClockWidthMode Mode, double CustomWidth) : SettingChange;
+    public sealed record ClockWidth(ClockWidthMode Mode, int CustomWidth) : SettingChange;
 }
